Remove pyright from pre-commit configuration
Pre-commit installs pyright in its own environment
without the project dependencies. This causes pyright
to complain about all modules being missing.
This can be fixed by either pointing pyright explicitly
to a venv (but that's annoying since the venvs for
the project are managed by hatch) or by explicitly
listing the project dependencies as `additional_dependencies`
in the pre-commit hook configuration (which is terrible
because it means maintaining two separate lists of project
dependencies).

At this point it's simpler to just have pyright
run in an editor and as a CI check.