Fix assertion in hot restart test
VMEC++ reports niter == 2 after a run that actually
only performs one iteration with this output:

```
 NS = 51   NO. FOURIER MODES = 59   FTOLV = 1.000e-06   NITER = 60000

 ITER |    FSQR     FSQZ     FSQL    |    fsqr     fsqz      fsql   |   DELT   |  RAX(v=0) |    W_MHD   |   <BETA>   |  <M>
------+------------------------------+------------------------------+----------+-----------+------------+------------+-------
    1 | 9.96e-07  1.35e-07  4.69e-09 | 9.35e-12  2.49e-12  9.85e-11 | 5.00e-01 | 8.789e-01 | 4.4971e-02 | 0.0000e+00 | 1.379
MHD Energy = 4.497072e-02

NUMBER OF JACOBIAN RESETS = 0
```

We might want to fix the niter behavior independently, this commit
is a quick fix for the broken test.

The reason we did not notice is that these tests do not actually
run in the CI (fixed in the next commit).