BaseModelWithNumpy for numpy+pydantic serialization (#264)
### TL;DR

Added a class to help with serialization and deserialization of classes with numpy arrays annotated with jaxtyping.

Differences to serialization in repo:
- No binary blobs
- Works with `jaxtyping > 0.3.0` by matching on `jt.AbstractArray`

## Why?
numpydantic is aiming to solve a lot more than we need and caused more issues with the type system than it solved 😥.
Internally we use a different approach, which works nicely. This stripped down version of that library should make integration smoother.