Skip gradual vacuum pressure activation on hot-restart (#330)
Usually VMEC turns on the pressure gradually over the first few iterations, to improve robustness. When hot-restarting, we don't want to do that, because the initial guess should already be close to the final solution. Therefore we initialize the vacuum state to `kInitialized` from the beginning, which transitions to `kActive` after the first iteration.

One additional change to the iteration logic is required for everything to work:

The problem is that only the pre-conditioned force-balance includes the vacuum contribution, but the termination criterion (`fsqr<ftol && fsqz<ftol && fsql<ftol`) depends only on the regular, invariant forces. This means a hot-restarted, free-boundary iteration would immediately, if the inner surfaces are in force-balance, eventhough the vacuum force-balance at the LCFS may not be fullfilled at all!
Additional logic is required to handle that case, and prevent termination at 0 iterations.