Fix a race condition on iter2. (#387)
iter2_ is read in Printout and updated in all threads after that read.
However, there seemed to be an overlap of incrementing it with printing
it out. This PR prevents this by using a backup of iter2 from the
beginning of the force iteration, which seems to solve that data race
(at least according to TSan/Archer and screen output).