szip requirement in netcdf, limit version (#397)
**This fixes the pip wheel failiure**

The new NetCDF version made szip a hard requirement. Since we ship binary dependencies with the wheel, we decided to limit the netcdf version for now instead of shipping a larger binary package.

This is not a permanent solution, and if it becomes a problem, we need to add szip to the linux images for the wheel build.